requirements.txt
requests
beautifulsoup4
markdownify
ftfy
langchain-text-splitters
langchain-core
langchain-ollama
langchain-groq
langchain-chroma
chromadb
python-dotenv
pydantic-settings
fastapi 
uvicorn[standard]
pymupdf